Move `O.C.LedgerDB.HD.*` modules.
The `HD` directory doesn't exist anymore in `ouroboros-consensus`.
