+++ ouroboros-consensus/test/consensus-test/Main.hs
@@ -4,6 +4,7 @@ import qualified Test.Consensus.BlockchainTime.Simple (tests)
 import qualified Test.Consensus.HardFork.Forecast (tests)
 import qualified Test.Consensus.HardFork.History (tests)
 import qualified Test.Consensus.HardFork.Summary (tests)
+import qualified Test.Consensus.Ledger.Tables.DiffSeq (tests)
 import qualified Test.Consensus.Mempool (tests)
 import qualified Test.Consensus.Mempool.Fairness (tests)
 import qualified Test.Consensus.Mempool.StateMachine (tests)
@@ -46,4 +47,5 @@ tests =
             Test.Consensus.HardFork.Forecast.tests
           ]
       ]
+  , Test.Consensus.Ledger.Tables.DiffSeq.tests
   ]
