+++ ouroboros-consensus/test/storage-test/Test/Ouroboros/Storage/LedgerDB.hs
@@ -7,15 +7,15 @@
 module Test.Ouroboros.Storage.LedgerDB (tests) where
 
 import           System.Info (os)
+import qualified Test.Ouroboros.Storage.LedgerDB.BackingStore as BackingStore
 import qualified Test.Ouroboros.Storage.LedgerDB.DbChangelog as DbChangelog
 import qualified Test.Ouroboros.Storage.LedgerDB.DiskPolicy as DiskPolicy
-import qualified Test.Ouroboros.Storage.LedgerDB.HD as HD
 import qualified Test.Ouroboros.Storage.LedgerDB.InMemory as InMemory
 import qualified Test.Ouroboros.Storage.LedgerDB.OnDisk as OnDisk
 import           Test.Tasty
 
 tests :: TestTree
-tests = testGroup "LedgerDB" ([ HD.tests | os /= "mingw32" ] <>  -- FIXME: Should re-enable at some point, see #4022
+tests = testGroup "LedgerDB" ([ BackingStore.tests | os /= "mingw32" ] <>  -- FIXME: Should re-enable at some point, see #4022
                               [ InMemory.tests
                               , OnDisk.tests
                               , DiskPolicy.tests
